requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
transformers==4.35.2
torch==2.1.1
pandas==2.1.3
numpy==1.26.2
python-multipart==0.0.6
pydantic==2.5.2
sqlalchemy==2.0.23 